language-specific theme support for ultraviolet
Some ultraviolet themes are better at highlighting some languages.  You
may now specify which theme to use depending on the language.  Simply
add a :themes argument to your config.ru use line that looks something
like this:

    :themes => {"vibrant_ink" => ["ruby"], "upstream_sunburst" =>
      ["objective-c", "java"]}

It is a hash whose keys are theme names and whose values are arrays of
language names.

If there is no match, :theme is used as the default.