Support Transfer-Encoding with Phusion Passenger
Passenger sets a Transfer-Encoding chunked header and only checks for the specific 'Content-Length' case to strip out a content length header. Although this could be said it’s a bug in passenger it’s also a symptom of using the Rack::Utils::HeaderHash which provides case insensitive readers but retain case sensitivity when assigning.